Tentative fix for out-of-bound scroll detection on named cursors
It doesn't work.

Postgres cursors are base-1, while Python (at least the scroll()
implementation for client-side cursors) is base 0. Accidentally the two
seem working the same, because MOVE is actually a FETCH without result,
so MOVE 4 is "go on the 4th record and discard it", hence the cursor
ends up on the 5th record, which is what expected with a Python cur[4].

The magic breaks trying to move on the first record.  MOVE ABSOLUTE 0,
works ok, but it is actually an OOB that leaves the cursor on the first
record.  That's still ok: we can special-case `scroll(0, 'absolute')` to
not raise an error even if the statusmessage is 'MOVE 0'.

What we cannot tell apart is if `scroll(-1)` went OOB or not. If we are
on the 2nd record is fine moving -1, but if we are on the 1st record,
going -1 is an oob. In both cases the statusmessage is 'MOVE 0' and we
can't tell them apart. In this case 'MOVE -1' should be allowed::

    piro=> declare tmp scroll cursor for select * from generate_series(0,10);
    DECLARE CURSOR
    piro=> move absolute 1 from tmp;
    MOVE 1
    piro=> move -1 from tmp;
    MOVE 0
    piro=> fetch 1 from tmp;
     generate_series
    -----------------
                   0
    (1 row)

But in this case not::

    piro=> move absolute 0 from tmp;
    MOVE 0
    piro=> move -1 from tmp;
    MOVE 0
    piro=> fetch 1 from tmp;
     generate_series
    -----------------
                   0
    (1 row)

but the result is 'MOVE 0' in both cases.